+++ pkg/core/resources/apis/mesh/traffic_permission.go
@@ -44,15 +44,15 @@ func (t *TrafficPermissionResource) SetSpec(spec model.ResourceSpec) error {
 var _ model.ResourceList = &TrafficPermissionResourceList{}
 
 type TrafficPermissionResourceList struct {
-	Total      uint64
+	Total      uint32
 	Items      []*TrafficPermissionResource
 	Pagination model.Pagination
 }
 
-func (l *TrafficPermissionResourceList) GetTotal() uint64 {
+func (l *TrafficPermissionResourceList) GetTotal() uint32 {
 	return l.Total
 }
-func (l *TrafficPermissionResourceList) SetTotal(total uint64) {
+func (l *TrafficPermissionResourceList) SetTotal(total uint32) {
 	l.Total = total
 }
 func (l *TrafficPermissionResourceList) GetItems() []model.Resource {
